requirements.txt
discord.py==2.3.2
psycopg2-binary==2.9.7
retry==0.9.2

